Add files via upload
Assignment Overview
After cleaning the dataset, your next step is the analysis. In this stage you will become more familiar with the data set and it will start to take shape. In this assignment you will:

Plot a distribution curve, and histogram.

Find the median, and outliers of particular columns.

Compute the Inter Quartile Range.

Find out the upper and lower bounds, and find correlations between numerical columns.

Create a new dataframe.